Add Character::read_time_multiplier()
This calculates a multiplier to book reading times which factors in
intelligence, traits, mutations, CBMs, etc... In that sense, it can be
interpreted as a Character's innate reading ability regardless of
external factors. Added new tests for it.

Character::read_speed() is rewritten to use this new function. The
read_speed tests passed.